Rewrote the rate-limiting code
I wasn't sure why the previous rate-limiting code didn't work so
I basically rewrote it using a new strategy.

If you set sendsPerSec to 4, it will ensure that there is a delay
of 250ms between each command sent.

I tested this with my tagbot for a bit and even with 20+ search
results it didn't get flood kicked.