Moved logging of the response objects into _onResponseReceived
I moved the code that logs the incoming response objects so that
the responses are logged before the response events are emitted.
Before, it was handling the response events first, then logging
that it received the response, which lead to very confusing log
messages.

For example, the PONG response was being logged before the PING
that it was responding to.